swap tblib dependency for native RaytaskError
