+++ python/ray/serve/utils.py
@@ -7,6 +7,7 @@ import string
 import time
 from typing import Iterable, Tuple
 import os
+import traceback
 
 import requests
 import numpy as np
@@ -14,6 +15,7 @@ import pydantic
 
 import ray
 import ray.serialization_addons
+from ray.exceptions import RayTaskError
 from ray.util.serialization import StandaloneSerializationContext
 from ray.serve.constants import HTTP_PROXY_TIMEOUT
 from ray.serve.http_util import build_starlette_request, HTTPRequestWrapper
@@ -221,3 +223,16 @@ def ensure_serialization_context():
     been started."""
     ctx = StandaloneSerializationContext()
     ray.serialization_addons.apply(ctx)
+
+
+def wrap_to_ray_error(function_name: str,
+                      exception: Exception) -> RayTaskError:
+    """Utility method to wrap exceptions in user code."""
+
+    try:
+        # Raise and catch so we can access traceback.format_exc()
+        raise exception
+    except Exception as e:
+        traceback_str = ray._private.utils.format_error_message(
+            traceback.format_exc())
+        return ray.exceptions.RayTaskError(function_name, traceback_str, e)
